requirements.txt
python-telegram-bot==20.7
pandas==2.2.2